Upgrade actions/checkout@2 -> actions/checkout@3
We are getting warnings in the output from the CI pipelines stating that
Node 12 which the current action uses is being phased out:

"Node.js 12 actions are deprecated. Please update the following actions to use Node.js 16: actions/checkout@v2. For more information see: https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/."