Change to session fixtures instead of module
The tests doesn't seem to get in each others way and recreating the
environment takes up quite a lot of time as a percentage of the test run
for the SDK integration tests.

Can we run these fixtures as session fixtures instead of module ones?

Looks like it lowers the test time for the SDK integration test from
80-90 seconds to 60-65 seconds.