Add TestHelper fixture
* Introduces a TestHelper class where we can add functions that can be
  relevant within the testing framwork.

I noticed this when working on #326 but didn't want to change something
unrelated in that PR.

It seemed a bit redundant to have a separate fixture to load each of the
schema files. I think it's quite handy to have a test class where we can
add other functions when needed.