Add way to replace problematic data in graphql mutations
Fixes #351

This PR replaces #360 and handles the replacement of variables before it
hits the Graphql objects.

In another iteration after the next release I think we can revisit the
top-level _generate_input_data and have it return an object instead of a
dict. I think that would be a bit cleaner. But will move back to that
later.