Add ruff to invoke linting tasks
Later we can start to phase out some of the other linters. For now this
will still catch what Ruff would find but the others not and it takes
less than 0.1 seconds to run through the entire codebase so it shouldn't
add any overhead.

Also added the path context to Backend and CTL so it's possible to run
these tasks from any location within the repo.