Configure gunicorn logging to play nice with structlog
We also needed to configure Uvicorn when running from within Gunicorn.
When running in devel mode with the infrahub server command we provide a
log config to uvicorn but it didn't seem to be possible to handle in the
same way in this scenario. As the gunicorn --logger-class requires the
path to an actual Python module as opposed to how the --config parameter
works I decided to move all related file into a separate directory.

Fixes #478