Consolidate error handling
Changes the main exception handling so that we only use the base Error
for Infrahub to handle uncaught exceptions. Also changes the format of
the error data so that we can reuse the same errors for both the REST
API and the GraphQL endpoint. For the graphql endpoint we also replace
the original HTTP status code with 200.

This will be used to handle authentication errors in the same way
regardless of if the request is to REST or GraphQL.