Dynamically select RabbitMQ port based on TLS setting
This is in preparation for #496. Once we have encrypted traffic between
the different components it will not make sense to have to modify both
the option to enable TLS and also change ports.

The PR sets sensible defaults for the port that are changed depending on
the status of TLS. If a user sets the port to a value that would
override the defaults.