Add code to enforce authentication on change operations
* Modifies the response to auth requests to return a user session
  object that contain the permissions
* Adds two new options for the configuration one to allow anonymous read
  access and one to ignore authentication (will be removed when the
  frontend and SDK supports authentication)
* Modifies the current auth middleware backend with a slight change of
  format
* Adds "role" to account and assigns the default admin user to the role
  of "admin"

For now there are no added tests for the authentication (aside from the
changed one to fix the format), this is mainly due to the config option
to ignore authentication. Next week I'll look at changing the tests so
that those are authenticated and review the test coverage of the
authentication in general.