Enable authentication requirements for demo
* Introduced a new config param to add an initial token for the admin
  user
* Adds missing headers for schema in the client
* Replaces `api_token` in the client with a Config object, will later
  also be used for #577, making it possible to specify SDK config using
  environment variables (we should move more of the parameters into this
  one later)
* Changes the Docker Compose file to add and use a new admin token (for
  now we allow a user to specify the admin token, we'll have to change
  this later)
* Change the Docker Compose file to enforce authentication for change
  operations
* Fix typing issues seen from first_initialization with regards to
  default values

Note, for now the initial token is hard coded in the docker compose
file. Once we change it so that the token is generated we'll have to
modify the workflow a bit so that users testing this needs to store the
token somewhere so it's made available to the git-agent when connecting
to the api.