Move data validation checks into git-agent
This PR moves the data validation checks so that they happen in the
git-agent instead of within the API. To be able to do this we had to
allow the git-agent to have direct database access.

In the current implementation there's one problem with that which is the
lack of a background job to update the schema like we have in the API. I
didn't look at it at all for this PR, mainly because it's not really
related and I don't think we'll have a problem as all of the nodes we
are interested in comes from the core schema and we only modify them in
the global branch. However we should probably add it later.

I added the database access to the service object. I would have wanted
to make it more generic but we can return to that part later.

Also I've just copied the code as is from the API and there are a few
improvements I'd like to make in later iterations.

1. It still behaves in the same way where it doesn't set the checks to
   be in progress before starting and instead just mimics what it did
   before. I didn't want to introduce any changes to the logic and think
   we should first confirm that we agree upon the overall approach for
   the database access within the git-agent
2. I want to add additional messages for this and would like that we try
   to stick to a rule that we only have one transaction per message,
   after that we should send a new message for further instructions. The
   reason for this is that it can be hard to recover to the same state
   if a message fails after having completed two out of three
   transactions. At this point we'll want to reschedule the message to
   try again later(*), and when we do we don't want it to perform the
   first two transactions again.

(*) For the retry parts I want to add some logic to that within the Meta
object of each message. As part of that I'd also want to remove the
Python pickle based messages and replace them all together with this
json format. I think it will be easier to keep track of each message
type and also easier to troubleshot within RabbitMQ if we have a json
payload, and of course if we add non-python agents.