Remove artifact tests that were duplicates of transform tests
As described in #976 the file
backend/tests/unit/git/test_git_artifact.py was just a copy of
backend/tests/unit/git/test_git_transform.py with one of the tests
renamed.

It didn't test artifacts at all and was as such missleading. Removing
the file for now to add proper tests later.