Add events to trigger actions in all workers based on changes
This PR adds some changes to how we work with RabbitMQ. The main
exchange where we publish JSON messages will get bound to the exclusive
queue that each worker now subscribes to meaning each worker (currently
4 gunicorn workers and 2 git-agents)

Most of the messages will only get handled by the git-agents, and the
responses to RPC requests will get handled by the gunicorn workers.

This new category bound by the routing key "refresh.registry.*" for
tasks related to refreshing the registry will be send to all workers. As
I figured it wasn't required for the initiating worker to refresh the
registry again I introduced an `initiator_id` to the message metadata
along with a variable infrahub.WORKER_IDENTITY which is just a uuid
unique to each of the workers at startup.

For the refresh I'm reusing the refresh_branches() method that
previously existed on the BackgroundRunner. I only moved it to
infrahub/tasks and changed the logger.

For now I didn't touch the BackgroundRunner aside from moving out the
method and instead calling an external function. After a while I think
we can increase the 10 second counter so that it doesn't run that often.

Fixes #1060