Switch to durable quorum queues and exchanges, add DLX
* Changes the queues and exchanges we want to live across restarts of
  the RabbitMQ server to be durable
* Changes the non-exclusive queues to be declared as quorum queues (not
  as relevant now but we want this in place in the future when we run
  in a clustered RabbitMQ environment)
* Adds internal routing within RabbitMQ to delay messages and a dead
  letter exchange where messages end up after the delay

This PR modifies the existing queues and exchanges in RabbitMQ, which
isn't possible. Due to this an `invoke demo.destroy` will have to be
issued in order to get the new settings to work.

This is something that we'll want to have in place to schedule checks
for the user defined repository checks.

It will also be useful for when we want to retry messages due to
failures. For example when implementing webhooks it could be that we
fail to reach the external service for the first try so we retry the
message with a delay.

This approach with dead letter exchanges was choosen in favor of the
Delayed Message Exchange plugin
(https://github.com/rabbitmq/rabbitmq-delayed-message-exchange), mainly
due to the fact that we'll always be able to see messages within a queue
whereas the delayed exchange hides them from view during the waiting
period.

A downside of the dead message approach is that we need to define a new
queue for each delay timer as the x-message-ttl is global for the queue.
Because of this the new Python class MessageTTL has been introduced,
we'll use this enum as an input parameter to the function where we want
to have a delayed message. If this optional param has a value we will
publish the message to the delayed queue and insert the expected `delay`
header so that the message is routed to the correct queue. Once the
timer has reached 0, the message gets rerouted to the final queue and
gets picked up by the git-worker.