New artifact messages and rewriting of API logic to create artifacts
This PR simplifies the API endpoint for generating artifacts. As the
user didn't get any actionable response it's currently replaced by just
an HTTP error code which would indicate that the requested artifact
definition was found in the system and that the request was sent to the
message_bus. Moving the logic to the worker also means that we can reuse
the same logic when a branch has been merged.

Related to #1070, though it doesn't look at the branch local aspects of
that issue.

Should also fix #1127 but I haven't verified that yet.

* event.branch.merge: Triggered when a branch is merged (sends trigger.artifact_definition.generate)
* trigger.artifact_definition.generate: Query all artifact definitions available on a branch and send request.artifact_definition.generate for each
* request.artifact.generate: Generates one artifact

* request.artifact_definition.generate: Previously sent an API request to generate artifacts based on the definition_id. Now instead sends a request.artifact.generate message for app applicable artifacts

Removed the last defined pickle based message.

There is still some cleanup to do before the old handlers can be deleted and the old
subscription code is still sending messages that are currently just
dropped by RabbitMQ as they aren't routable.

Also there's an old method on the repository to generate artifacts which
won't be used any more. There are still a number of unittests around
that method. In a follow up PR I will move the tests to the new method
instead and delete the previous one.