Switch default mutation messages to JSON
For now these messages doesn't do anything unless debug logging is
enabled. It's mostly a placeholder for the future but I want to get rid
of them in the current format. With this setup the data would only
include attributes and not relationships which would be similar to what
we had before as you'd need access to the database to fetch the
relationships. Once we actually need these messages we should probably
look at providing more information about what was actually modified on
the objects.