Merge branch when proposed change state is updated to merged
Also combines the update to the proposed change and merging of a branch
into a single transaction

Adds some checks to ensure that a proposed change that have been merged
or canceled can't be reopened or transition into another state.

As part of this I changed to use the get_one_by_id_or_default_filter as
that function raises an error if the node isn't found. This might be a
bit more costly as it can result in multiple queries if the object isn't
found. I didn't want to have to query for the object both in the
proposed_change update and in the generic update function. We could also
introduce a new function called something like .get_by_id_or_raise()
that would do the same.

Changed the frontend code so that it doesn't send in two mutations when
merging a proposed changed (one to merge the branch and one to update
the state of the proposed change)

This is related to #1211 but doesn't yet change the logic only ensures
that we can do the merge within one transaction and one mutation.