Initialize git-agent with services object
Make the services object available throughout the git-agent. The main
reason for this is to be able to initialize repositories and supply the
service object. This PR is related to the changes in #1429.

In this iteration the code to initialize RabbitMQ has just been copied
as they were before. A future step will be to have the RabbitMQ adapter
take care of this for both the git-agent and api-server and remove this
code from the git-agent all together.