Update schema for Python checks in .infrahub.yml
Add parameters and targets. If targets is empty this will be a global
check, otherwise it will target a specific group.

Also introduces InfrahubCheckInitializer, this is to provide information
to the check that user created code could act upon. For now it contains
information about which proposed change (if any) triggered the check to
run. The reason for this is to support that use case with checking
something within an external system such as ServiceNow to see if a given
proposed change has been approved. We don't yet inject the proposed
change into the check that will come in a follow up PR.

Related to #1049, will move out User checks from the
RepositoryValidators as a next step and also add the ability to target
groups of devices for these checks.