Cleanup of git-agent move rabbit logic to the adapter
Moves the initialization of the RabbitMQ code into the services
message_bus adapter for RabbitMQ. The thought is to both simplify the
setup within the CLI code and also to consolidate the setup code between
the Git agent and the API server. For now the code on the API server
remains intact.

At some later point I want to do the same for the API server and also
move the subscribe function into the RabbitMQ adapter and also move away
from the InfrahubRpcClient object.

Also introduces an initialize() method on the services object. I'd like
to eventually move the DB init and other setup type of of jobs so that
they work in the same way.