Fix ability to run checks from infrahubctl
The ability to run checks with infrahubctl has been broken for some
time. This PR repairs the functionality and also simplifies some of the
nested code. Also moves out the functions to load the .infrahub.yml file
from the repository into a separate file so that this functionality
isn't split between utils.py and cli.py.

This is in preparation for #1049.