Refactor initialization of RabbitMQ on API server
Moves the initialization into the RabbitMQ adapter. Adds an override
setting to the config to have an easier access to override the choice of
which adapter to use within the test framework instead of replacing
objects with a mock.

Delete some functions from InfrahubRpcClient that are no longer needed
there. The InfrahubRpcClient is still used in quite a few places and I
didn't address that part now as there were quite a few changes already.
In a future PR I'd like to remove InfrahubRpcClient &
InfrahubRpcClientTest alltogether and modify some tests so that we can
better test the workflow from API server to the tasks that would be
performed by the Git agents.