Convert Override class to dataclass
The reason for this is to avoid having to import the neo4j driver just
by reading the config. The override class is only used within testing
and when we want to support git-agents without database access we want
to lower the memory footprint on them as much as possible, currently the
neo4j driver is also imported using the git-credentials helper. There's
still some cleanup to avoid doing that. This provides part of it.