Modify error levels when logging GraphQL errors
Some errors are currently very noisy with regards to what is logged. If
a user requests a node that doesn't exist it shouldn't look like an
error occured in the server.

For how any of our own errors defined at HTTP level 500 will be logged
in the same way as before. For any permission denied error we log them
at the info level to not have them completely silent.

Other errors that are strictly user errors that we don't want to have in
the logs have been moved to the debug level for now. I'm sure we will
have to move them back and forth a bit to find the correct level for
each type.

Any unhandled exception that we didn't account for is treated as an
actual error.

Fixes #2318.