Split apart infrahub.core.schema into smaller parts
This is in preparation for #945. The schema.py file was over 2500 lines
long. With the current changes the schema/definitions/core.py file is
still over 1000 lines but overall this should be easier to manage.

In upcoming steps we will introduce a module that contains code that is
generated based on the Python dictionaries. Before we get there I wanted
to be in a state where we had better typing within those files which was
what started this change.