Adds initial task logs for request.proposed_change messages
Also adds verbosity for messages when they fail the last time for easier
troubleshooting. We will scale this down again later. Probably also need
to catch other type of errors before we get there so that we keep this
kind of verbosity for unhandled errors and otherwise print a friendly
message for errors that are expected.

For now I've mostly just indented the code in these messages to include
the most basic of Task report. Will add it to the following messages and
include more details in a later step.