Initial integrationtest for proposed change
Tests a successful and failing integration test from the API and reads
the status of the data integrity checks

Also includes some thoughts about structure for the test schema. This is
to avoid having to redefine it in multiple places. Potentially we could
run into issues if we want to modify these schema files in a way that
causes other tests to fail. On the other hand it's very easy to create
new schemas with this structure.

Also includes some minor typehint work. As I had to access properties in
the Node object I had to add some exceptions to my test :(