Add merging of proposed change to happy test and fix value bug
A bug was introduced in a recent PR adding an additional .value under an
existing value but it wasn't caught as we didn't have a test that
covered that path. The result is that currently merging a proposed
change is broken. This PR adds a test to avoid regressions in the
future.