fix(tests, e2e): flaky select-2-steps
If the InfraInterfaceL3 list query takes some time, it will prevent the
device dropdown to show when it completes.
We should either fix the frontend to properly show the dropdown in this
case or wait for the InfraInterfaceL3 query response before resuming the
test.

This is reproducible by inserting a sleep(2) in the graphql endpoint.

Signed-off-by: Fatih Acar <fatih@opsmill.com>