Fix bug related to choice of data conflict resolutions
Selecting the source branch as the branch to keep when resolving a data
conflict within a proposed change previously didn't work. This PR fixes
that behaviour and adds a test to validate that it works.

Fixes #2984

Also added another fix where we tried to merge git repositories for the
source branch of all connected repositories even though the source
branch wasn't setup to sync to git (i.e. previously is-data-only) This
caused the merge operation within the message bus to fail since the
source branch couldn't be found as it doesn't exist in git.