Fix path traversal vulnerability in the git agent (#2996)
The git agent currently does not validate the proper path when being
queried on its /api/file/{repo}/{file} endpoint. This allows an attacker
who knows about the name of a repo to crawl back to the root of the file
system, and can therefore read files, if permission allows.

The git agent must validate that a file is part of a specific directory
before trying to access and to read it.