Use HFID across the demo schema
There are some unique/optional values that were adjusted to apply HFID.

A part of the schema manager HIF validation logic has also been
commented to allow the use of non-unique values in the HFID as long as
one item in the HFID is guaranteed to be unique. This should be
revisited befoe merging.