Migrate exception handler to pydantic v2
Pydantic v2 seems to work different compared to v1 with regards to how
errors are raised. Since the test was using a made up field with a
custom error I instead opted to create a new field which I think would
be a closer match to what we see in production regardless.

Also made some minor changes to clean up typing in the tests.