Expose GraphQL search anywhere query (#3688)
The search is now using a dedicated query. The frontend code has also
been adapted to use this new query.

The search logic will try to look for a specific node if the search
parameter looks like a valid UUID. It will default to the previous
search behaviour in other cases.