Activate Ruff ANN rule for annotations
This is fairly similar to what mypy does. The main difference is that
ruff can find some of the issues faster. Aside from the speed when
developing the main benefit is that it can be easier to fix a sub
section of code and resolve some of the issues at a time. The violations
are ignored per section and can be fixed or split apart further as parts
of the code gets changed so that it's in compliance with the typing.