Make permission report branch aware
This will allow user to see permission decision according to what branch
they are in. For instance, if a user is in the default branch and has
the permission to do something, report will say "allow". In the same
case, but in a non-default branch, without the permission to do the same
thing, report will say "allow_default". Other cases are also covered.