fix(ci): update release workflow to use release events
This fixes issues with pre-releases where the git tag may not match the
previous filter we had.

We now check if the github release is a "pre-release" and also check if
the Poetry version is also a pre-release or dev-release one.

If the github release is not pre-release, we ensure that the Poetry
version is also not a pre-release one.

We also fix a potential issue where the "latest" Docker tag would be
added to older versions. Leverage the GitHub "latest" release flag to
solve that.

Signed-off-by: Fatih Acar <fatih@opsmill.com>