Introduce permission manager
This object is build per user session / per request and is carried over
in the GraphQL context. This allows to fetch permissions only once and
to perform lookups in memory.

The permission manager is in charge of loading permissions from the
configured backends. Therefore these backends are now simple constructs
that load permissions from a store before handing them over to the
manager.

The permission manager also takes care of deciding when to grant or
revoke a permission for a user. It also reports permissions for a user
when requested.