IFC-592 Merge schemas when loading them from files (#5482)
When loading multiple schemas files, they used to be imported one by
one. With this change, the files are now parsed as schemas and then
merged together in order to produce a single schema to load. The merge
is performed one by one in the order of the files. This means that the
schema coming last will overwrite the previous one if they refer to the
same nodes.