Make `parent` mandatory in some cases
The `parent` relationship becomes mandatory if the node depends on
another one which is not itself or a generic of itself. This allows, for
instance, to not require a continent to have a parent while enforcing a
country to have a parent continent. The parent won't be required if a
node can depends on one of its own. For instance, an IP prefix can have
a parent prefix, which is still optional as an prefix can live on its
own.