Revert proposed change state to open on merge failure
If we merge a proposed change and there's an error we previously didn't
revert the state back to "open" from "merging", this PR corrects this
behaviour. Previously we did revert this if a user tried to merge a
proposed change with data conflicts without providing a resolution to
all conflicts.

Fixes #5563