feat(backend): add database connection throttling
Also add metrics related to in use connection.
First attempt was to use metric callback so that we grab the actual
usage but was hard since we require access to the request context that has the
DB driver object referenced.
A simpler version is to keep track of the last usage value seen and
update the metric accordingly.

Note that the max_concurrent_queries setting should match the query
thread pool on the server side (defaults to 400) and be distributed
among server workers.

Signed-off-by: Fatih Acar <fatih@opsmill.com>