Fix HFID duplicates for component templates
When loading several schemas without a restart the logic was adding the
same HFID and uniqueness constraint element more than once for component
templates. This lead to a final schema with unpractical HFID and
uniqueness constraints (mostly from a user perspective). While a restart
would fix the issue, this change makes sure it does not happen between
restarts.

An integration test has also been added to highlight the issue in case
of a regression.