Remove refs to `InfrahubServices` for git ops
This change tries to remove the need for a `InfrahubServices` object
when it comes to operations to be performed on a git repository.

This work is a stepping stone to be able to run user flagged workflows
inside a dedicated prefect worker for isolation purpose as we won't be
able to inject a `InfrahubServices` object into, also we don't want to
for security reasons.