fix(backend): make sure filter subqueries start from the node
Without this change, the planner will use the peer node index (either
node uuid or attribute value index) as the starting point of the query.
Thus in the case of IPAM, when we have a single namespace with thousands
of addresses/prefixes, it will start the query from the namespace and
traverse all the addresses/prefixes inside that same namespace (thus
retraversing all addresses for each address...).

Unfortunately, it looks like there's no support to "ignore" an index, so
a little trick here is to filter the actual starting node using its own
uuid to force the planner to use that node as the starting point.

Signed-off-by: Fatih Acar <fatih@opsmill.com>