IFC-1517 Don't set unique constraints on generic templates (#6565)
When we run the uniqueness constraint checker for a given node, we check
for its uniqueness constraints as well as the uniqueness constraints of
its generics if their are any. In the case of component template, it was
causing an issue where component templates were expected to have unique
template names while in fact, their uniqueness is based on a name and a
parent template.