Add `common_parent` relationship attribute (#6626)
This relationship attribute is used to specify that peers of a given
relationship must share a common parent with the source object. While
the parent of the source object is infered automatically (as there is
only one parent per object), the `common_parent` value must be set to
the name of the parent relationship of the peer schema.