Avoid going through the whole prefix to find gaps
This commit improves the gap finding logic for both addresses and
sub-prefixes in order to only find gaps in a prefix within a border
given by 2 addresses or 2 prefixes.

The pagination and the count do not include the "available" nodes. This
implies that if a limit of 4 is set for the query, the count will be
lesser or equal to 4 even though the number of edges might be greater
than 4 because of the inclusion of available nodes.