IFC-1690 Fix available prefixes when parent prefix is empty (#6874)
This change makes sure that we return subnets of the container prefix
for which we asked the available space. When a container prefix has no
space reserved, the next 2 smaller subnets that fit inside the container
must be returned. For instance, if 192.0.2.0/24 has no subnets, the
available space query must report 192.0.2.0/25 and 192.0.2.128/25.