IFC-1539 Add query to get actions on a proposed change (#6926)
This change introduces a new `CoreProposedChangeAvailableActions` which can be used to understand what are the actions that can be taken on a proposed change. While the way the code is built can certainly be improved, this PR is mostly here to propose and expose an API that we should agree on.

The list of possible actions to take and how they are evaluated may change. If it comes to that it should be fairly simple to adapt.