Replace new_branch variable and ignore updates for -global- branch
The variable name was confusing as all of the branches weren't new,
switching to active_branch for clarity.

Due to a recent refactoring we were logging the warning message:

"Branch schema hash is not set, cannot update branch registry"

The root cause of this was that we didn't have a schema_hash on the
'-global-' branch which we never was supposed to have.