Revert "fix: make sure to lock on resource manager allocations"
BROKEN: this does not handle double lock/deadlock issues when we acquire
a resource pool lock in run_constraints_and_save() and then in
get_resource(). We probably require a mechanism to check if the lock is
actually held by the current worker, if so it can proceed without
acquiring the lock.

This reverts commit 31d79407eb8e0111c024693d7c3a7cec66e8bde2.