fix(backend): make enriched_nodes_create scale
Using a separate MERGE statement for the DiffRoot makes the subsequent
MERGE statement for the DiffNode not use the index on DiffNode and would
rather make the query start at the DiffRoot, thus degrading performance
when a lot of DiffNodes are attached.
Using a single MERGE statement makes the planner use the DiffNode uuid
index effectively.

When having large diffs, this can lower down last batch execution
duration from 10s to 2s.

Signed-off-by: Fatih Acar <fatih@opsmill.com>