fix(backend): cache git repo initialization
When generating a lot of artifacts at the same time (for example when
merging a huge branch). Repo initialization will be done for each
artifact, concurrently.
Avoid useless initialization steps that triggers many graphql queries to
fetch and update the repo status by leveraging Prefect caching with a
small TTL.

Signed-off-by: Fatih Acar <fatih@opsmill.com>