IFC-1812 Refactor HFID & display_labels
This brings several changes to how display_label and human_friendly_id for nodes are stored and retrieved. While it tries to change the least amount of code to keep code compatible as much as possible, it now stores the display_label and human_friendly_id of nodes into the database to be able to retrieve them when looking up nodes instead of computing them every times. It also introduces events to help in display_label computation.

This provides implementations for IFC-1820, IFC-1821, IFC-1822, IFC-1823, IFC-1824, IFC-1825 and IFC-1826.