Allow rebase of branches during migration (#7447)
These changes add a new flow to apply migrations (schema/data ones) inside branches. The upgrade process (via the provided command) is also updated with a new `--rebase-branches` flag to trigger branch rebases as part of the upgrade process.

A new `MigrationRequiringRebase` class is introduced with a couple of methods that need to be defined on classes inheriting it. One method defines what to do on the default branch while the other one defines what to do for all other branches.

The branch nodes can have a new `NEED_UPGRADE_REBASE` status and they also now have a graph_version property to know if a branch is up-to-date regarding the graph structure.

A new prefect event is also introduced and triggered when the `branch-migrate` flow completes for a branch.