IFC-1984: Update git sync to consider new configuration (#7486)
* IFC-1983: Add config parameter for selective git sync

This PR adds `sync_branch_names` parameter to git settings so that users can add branches that they want to be synced with git
instead of syncing all branches like we currently do.
This parameter can also be a regex so that we can match branch names when syncing

* update docs

* update doc text

* IFC-1984: Update git sync to consider new configuration

This PR updates the git sync to check the new configuration `sync_branch_names` and if it is present
it matches the branch name with the names in the config skips adding it as a remote branch if it does not match.

* add configuration check on creating a branch

* update branch creation validation condition

* update validation condition

* move validation function outside model

* sync branch names compiles to regex, remove compiled branch names

* remove compiled branch names

* update git setting name

* remove main in description

* update git setting name

* refactor remote branches filter

* fix mypy

* rename function

* use sdk instead of client

* update sdk

* refactor get_filtered_remote_branches

* add changelog