IFC-1952 Add generic/attribute relationships in profile schema
This change is relatively simple and just adds relationships of kind generic and kind attribute to profile schemas corresponding to node schemas having these relationships.

When we load a schema, we make sure to clear all the relationships, except the ones of kind PROFILE such as related_nodes, to create a brand new list of relationships which will account for deleted and new ones. This is very similar to what we do for templates already.