fix(backend): cache task count results
When the task history is large, the count query can take a few seconds
to complete due to how the Prefect database is handling the query (it
will scan each row in the table).
We can cache the count result when the history is larger than a
statically defined 100k entries.

Also update the default db timeout value for Prefect since the count
query can sometimes take more time than the default 10 seconds value.

Fixes #7568

Signed-off-by: Fatih Acar <fatih@opsmill.com>